requirements.txt
Flask
python-binance
pandas
gunicorn